Rename cgreen/constraint_syntax_helpers.h to cgreen/constraints.h
Rename the file to `cgreen/constraints.h` so that it is clear this is the
file that should included if we need to use constraints when doing
tests.
If the objective is to create user defined constraints the file that
should be included is `cgreen/constraint.h`.
