+++ include/cgreen/cgreen.h
@@ -5,7 +5,7 @@
 #include <cgreen/cdash_reporter.h>
 #include <cgreen/cute_reporter.h>
 #include <cgreen/assertions.h>
-#include <cgreen/constraint_syntax_helpers.h>
+#include <cgreen/constraints.h>
 #include <cgreen/runner.h>
 #include <cgreen/boxed_double.h>
