+++ src/CMakeLists.txt
@@ -59,7 +59,7 @@ set(cgreen_SRCS
   cgreen_time.c
   cgreen_value.c
   constraint.c
-  constraint_syntax_helpers.c
+  constraints.c
   cute_reporter.c
   cdash_reporter.c
   messaging.c
